program_runtime: move cache entry logic to its own file
