+++ programs/bpf_loader/src/lib.rs
@@ -13,7 +13,7 @@ use {
     solana_program_runtime::{
         deploy_program,
         invoke_context::InvokeContext,
-        loaded_programs::{ProgramCacheEntry, ProgramCacheEntryOwner, ProgramCacheEntryType},
+        program_cache_entry::{ProgramCacheEntry, ProgramCacheEntryOwner, ProgramCacheEntryType},
         sysvar_cache::get_sysvar_with_account_check,
         vm::execute,
     },
@@ -982,18 +982,15 @@ fn common_close_account(
 
 #[cfg_attr(feature = "svm-internal", qualifiers(pub))]
 mod test_utils {
+    use solana_program_runtime::program_cache_entry::DELAY_VISIBILITY_SLOT_OFFSET;
     #[cfg(all(feature = "svm-internal", feature = "metrics"))]
     use solana_program_runtime::program_metrics::LoadProgramMetrics;
     #[cfg(feature = "svm-internal")]
     use {
-        super::*,
-        solana_account::ReadableAccount,
+        super::*, solana_account::ReadableAccount,
         solana_loader_v4_interface::state::LoaderV4State,
-        solana_program_runtime::loaded_programs::{
-            DELAY_VISIBILITY_SLOT_OFFSET, ProgramRuntimeEnvironment,
-        },
-        solana_sdk_ids::loader_v4,
-        solana_syscalls::create_program_runtime_environment,
+        solana_program_runtime::loaded_programs::ProgramRuntimeEnvironment,
+        solana_sdk_ids::loader_v4, solana_syscalls::create_program_runtime_environment,
     };
 
     #[cfg(feature = "svm-internal")]
