+++ svm/src/transaction_execution_result.rs
@@ -1,7 +1,7 @@
 use {
     crate::account_loader::LoadedTransaction,
     solana_message::inner_instruction::InnerInstructionsList,
-    solana_program_runtime::loaded_programs::ProgramCacheEntry,
+    solana_program_runtime::program_cache_entry::ProgramCacheEntry,
     solana_pubkey::Pubkey,
     solana_transaction_context::transaction::TransactionReturnData,
     solana_transaction_error::TransactionResult,
