+++ svm/tests/concurrent_tests.rs
@@ -14,9 +14,8 @@ use {
     solana_instruction::{AccountMeta, Instruction},
     solana_program_runtime::{
         execution_budget::SVMTransactionExecutionAndFeeBudgetLimits,
-        loaded_programs::{
-            ProgramCacheEntryType, ProgramCacheForTxBatch, ProgramRuntimeEnvironments,
-        },
+        loaded_programs::{ProgramCacheForTxBatch, ProgramRuntimeEnvironments},
+        program_cache_entry::ProgramCacheEntryType,
     },
     solana_pubkey::Pubkey,
     solana_svm::{
