+++ svm/tests/mock_bank.rs
@@ -11,7 +11,8 @@ use {
     solana_program_runtime::{
         execution_budget::{SVMTransactionExecutionBudget, SVMTransactionExecutionCost},
         invoke_context::InvokeContext,
-        loaded_programs::{BlockRelation, ForkGraph, ProgramCacheEntry, ProgramRuntimeEnvironment},
+        loaded_programs::{BlockRelation, ForkGraph, ProgramRuntimeEnvironment},
+        program_cache_entry::ProgramCacheEntry,
         solana_sbpf::{
             program::{BuiltinFunctionDefinition, BuiltinProgram, SBPFVersion},
             vm::Config,
